+++ main.go
@@ -1,15 +1,18 @@
 package main
 
 import (
-	"fmt"
 	"log"
 	"os"
 	"os/signal"
 	"syscall"
 
-	"github.com/granitebps/puasa-sunnah-api/middleware"
-	"github.com/granitebps/puasa-sunnah-api/routes"
-	"github.com/joho/godotenv"
+	config "github.com/granitebps/puasa-sunnah-api/configs"
+	"github.com/granitebps/puasa-sunnah-api/pkg/constants"
+	"github.com/granitebps/puasa-sunnah-api/pkg/core"
+	"github.com/granitebps/puasa-sunnah-api/route"
+	"github.com/granitebps/puasa-sunnah-api/scheduler"
+	"github.com/granitebps/puasa-sunnah-api/src/middleware"
+	"github.com/spf13/viper"
 
 	"github.com/gofiber/fiber/v2"
 )
@@ -20,21 +23,34 @@ import (
 // @contact.email granitebagas28@gmail.com
 // @license.name MIT
 // @BasePath /
+// @version 1.0
 func main() {
-	// Load ENV
-	err := godotenv.Load()
-	if err != nil {
-		log.Panic("Error loading .env file")
-	}
-	PORT := os.Getenv("PORT")
+	// Load ENV and setup some config
+	config.SetupConfig(".env")
 
-	app := fiber.New()
+	// Initiate Fiber
+	app := fiber.New(config.FiberConfig())
 
-	// Initialize Middlewares
-	middleware.InitMiddleware(app)
+	// Setup core package
+	conf := core.SetupCore()
 
-	// Initialize Routes
-	routes.InitRoutes(app)
+	// Setup middleware
+	middleware.SetupMiddleware(app, conf)
+
+	// Setup Dependency Injection
+	contr := SetupDependencies(conf)
+
+	// Setup route
+	route.SetupRoute(app, contr)
+
+	// Setup scheduler
+	scheduler.SetupScheduler(conf)
+
+	startServerWithGracefulShutdown(app)
+}
+
+func startServerWithGracefulShutdown(app *fiber.App) {
+	PORT := viper.GetString(constants.APP_PORT)
 
 	// Listen from a different goroutine
 	go func() {
@@ -46,14 +62,14 @@ func main() {
 	c := make(chan os.Signal, 1)                    // Create channel to signify a signal being sent
 	signal.Notify(c, os.Interrupt, syscall.SIGTERM) // When an interrupt or termination signal is sent, notify the channel
 
-	_ = <-c // This blocks the main thread until an interrupt is received
-	fmt.Println("Gracefully shutting down...")
+	<-c // This blocks the main thread until an interrupt is received
+	log.Println("Gracefully shutting down...")
 	_ = app.Shutdown()
 
-	fmt.Println("Running cleanup tasks...")
+	log.Println("Running cleanup tasks...")
 
 	// Your cleanup tasks go here
 	// db.Close()
 	// redisConn.Close()
-	fmt.Println("Fiber was successful shutdown.")
+	log.Println("Fiber was successful shutdown.")
 }