+++ pkg/constants/constant.go
@@ -1,0 +1,13 @@
+package constants
+
+const (
+	PRODUCTION  = "production"
+	DEVELOPMENT = "development"
+	LOCAL       = "local"
+)
+
+const (
+	FORMAT_DATE     = "2006-01-02"
+	FORMAT_DATETIME = "2006-01-02 15:04:05"
+	LOG_FOLDER      = "storage/logs"
+)