+++ CHANGELOG.md
@@ -8,6 +8,18 @@ You can find and compare releases at the GitHub release page.
 
 ## [Unreleased]
 
+## [0.1.0] - 2022-05-03
+
+### Added
+- Add category and type data to fasting response API
+- Add filter to fasting response API
+- Add sentry
+- Add testing
+- Add CI
+
+### Changed
+- Improve security
+
 ## [0.0.2] - 2022-04-20
 
 ### Added
@@ -32,5 +44,6 @@ You can find and compare releases at the GitHub release page.
 - Prepare for deployment
 
 [Unreleased]: https://github.com/granitebps/puasa-sunnah-api/compare/main...dev
+[0.1.0]: https://github.com/granitebps/puasa-sunnah-api/compare/v0.0.2...v0.1.0
 [0.0.2]: https://github.com/granitebps/puasa-sunnah-api/compare/v0.0.1...v0.0.2
 [0.0.1]: https://github.com/granitebps/puasa-sunnah-api/releases/tag/v0.0.1